// src/modules/carInfo/reduxDuck/types.ts
export interface Vehicle {
  make: string // the make of the car
  model: string // the model of the car
  enginePowerPS: number // engine power in Horsepower units
  enginePowerKW: number // engine power in KiloWatts
  fuelType: string // fuel type
  bodyType: string // body type
  engineCapacity: number // engine capacity in cc
}